More cleaning (#246)
* Remove unused font renderer

* Remove unused gui stuff

* Remove deprecated recipes for parametrizers and change creative tab item to high power casing. More iconic for the mod imo.

* Remove commands that are only used for a deprecated block

* Remove some unused config option for disabling GT material stuff

* TecTech chunk custom saving thing? Not entirely sure why this exists, it seems to serve no purpose?

* Clean up some unused code

* Remove a bunch of random stuff

* Seems unused?

* Spotless

* More cleaning + spotless, random unused stuff

* No swearing

* Update buildscript

---------

+++ src/main/java/com/github/technus/tectech/mechanics/tesla/ITeslaConnectable.java
@@ -117,12 +117,12 @@ public interface ITeslaConnectable extends ITeslaConnectableSimple {
                     if (origin.isOverdriveEnabled()) {
                         outputVoltageInjectable = origin.getTeslaOutputVoltage();
                         outputVoltageConsumption = origin.getTeslaOutputVoltage()
-                                + (distance * origin.getTeslaEnergyLossPerBlock())
+                                + ((long) distance * origin.getTeslaEnergyLossPerBlock())
                                 + (long) Math.round(
                                         origin.getTeslaOutputVoltage() * origin.getTeslaOverdriveLossCoefficient());
                     } else {
                         outputVoltageInjectable = origin.getTeslaOutputVoltage()
-                                - (distance * origin.getTeslaEnergyLossPerBlock());
+                                - ((long) distance * origin.getTeslaEnergyLossPerBlock());
                         outputVoltageConsumption = origin.getTeslaOutputVoltage();
                     }
